+++ gsa/src/web/entity/icon/trashicon.js
@@ -24,7 +24,7 @@ import {isDefined} from 'gmp/utils/identity';
 import PropTypes from 'web/utils/proptypes';
 import withCapabilities from 'web/utils/withCapabilities';
 
-import TrashIcon from 'web/components/icon/trashicon';
+import TrashcanIcon from 'web/components/icon/trashcanicon';
 import {getEntityType, typeName} from 'gmp/utils/entitytype';
 import {useCapabilities} from 'web/utils/useCapabilities';
 
@@ -67,7 +67,7 @@ const EntityTrashIcon = ({
     }
   }
   return (
-    <TrashIcon
+    <TrashcanIcon
       {...props}
       title={title}
       value={entity}