+++ gsa/src/web/pages/tasks/icons/resumeicon.js
@@ -22,19 +22,24 @@ import _ from 'gmp/locale';
 import {isDefined} from 'gmp/utils/identity';
 
 import PropTypes from 'web/utils/proptypes';
+import {capitalizeFirstLetter} from 'gmp/utils/string';
 import withCapabilities from 'web/utils/withCapabilities';
-import {useCapabilities} from 'web/utils/useCapabilities';
 import ResumeIcon from 'web/components/icon/resumeicon';
 
-const TaskResumeIcon = ({task, onClick, ...props}) => {
-  const capabilities = useCapabilities(props.capabilities);
-
+const TaskResumeIcon = ({
+  capabilities,
+  task,
+  usageType = _('task'),
+  onClick,
+}) => {
   if (task.isContainer()) {
     return (
       <ResumeIcon
         active={false}
         alt={_('Resume')}
-        title={_('Task is a container')}
+        title={_('{{usageType}} is a container', {
+          usageType: capitalizeFirstLetter(usageType),
+        })}
       />
     );
   }
@@ -44,7 +49,9 @@ const TaskResumeIcon = ({task, onClick, ...props}) => {
       <ResumeIcon
         active={false}
         alt={_('Resume')}
-        title={_('Task is scheduled')}
+        title={_('{{usageType}} is scheduled', {
+          usageType: capitalizeFirstLetter(usageType),
+        })}
       />
     );
   }
@@ -60,7 +67,7 @@ const TaskResumeIcon = ({task, onClick, ...props}) => {
       <ResumeIcon
         active={false}
         alt={_('Resume')}
-        title={_('Permission to resume task denied')}
+        title={_('Permission to resume {{usageType}} denied', {usageType})}
       />
     );
   }
@@ -69,7 +76,9 @@ const TaskResumeIcon = ({task, onClick, ...props}) => {
     <ResumeIcon
       active={false}
       alt={_('Resume')}
-      title={_('Task is not stopped')}
+      title={_('{{usageType}} is not stopped', {
+        usageType: capitalizeFirstLetter(usageType),
+      })}
     />
   );
 };
@@ -77,6 +86,7 @@ const TaskResumeIcon = ({task, onClick, ...props}) => {
 TaskResumeIcon.propTypes = {
   capabilities: PropTypes.capabilities.isRequired,
   task: PropTypes.model.isRequired,
+  usageType: PropTypes.string,
   onClick: PropTypes.func,
 };